requirements.txt
flask
flask-sqlalchemy
gunicorn
python-dotenv
pybit
ciso8601
mysqlclient
python-dateutil